Added feature dependency to menu items.
