+++ src/Abp/Application/Navigation/MenuItemDefinition.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Abp.Application.Features;
 using Abp.Collections.Extensions;
 using Abp.Localization;
 
@@ -43,6 +44,12 @@ namespace Abp.Application.Navigation
         public string RequiredPermissionName { get; set; }
 
         /// <summary>
+        /// A permission name. Only users that has this permission can see this menu item.
+        /// Optional.
+        /// </summary>
+        public IFeatureDependency FeatureDependency { get; set; }
+
+        /// <summary>
         /// This can be set to true if only authenticated users should see this menu item.
         /// </summary>
         public bool RequiresAuthentication { get; set; }
@@ -76,7 +83,8 @@ namespace Abp.Application.Navigation
             bool requiresAuthentication = false, 
             string requiredPermissionName = null, 
             int order = 0, 
-            object customData = null)
+            object customData = null,
+            IFeatureDependency featureDependency = null)
         {
             if (string.IsNullOrEmpty(name))
             {
@@ -96,6 +104,7 @@ namespace Abp.Application.Navigation
             RequiredPermissionName = requiredPermissionName;
             Order = order;
             CustomData = customData;
+            FeatureDependency = featureDependency;
 
             Items = new List<MenuItemDefinition>();
         }
